修改估值返回类型为int；调整连五权重以防溢出

+++ renju/cutoffTest.h
@@ -5,7 +5,7 @@
 
 #define MAX_DEPTH_FIXED 3
 
-inline bool cutoffTest(const GameFullStatus &status, const int &depth, const double &alpha, const double &beta) {
+inline bool cutoffTest(const GameFullStatus &status, const int &depth, const int &alpha, const int &beta) {
     return depth >= MAX_DEPTH_FIXED;
 }
 #endif //RENJU_CUTOFFTEST_H
